+ new algorithm for altrenate sequence construction implemented (the reference sequence is always a linear DB graph (vertices representing the same kmers contains additional suffix to be distinguished)  + some bugfixes (thread safety, memory lekas)

+++ kgraph/kgraph.c
@@ -174,7 +174,7 @@ int main(int argc, char *argv[])
 
 							ret = input_refseq_to_regions(seq, len, &regions, &regionCount);
 							if (ret == ERR_SUCCESS) {
-								uint32_t regionIndex = 0;
+								size_t regionIndex = 0;
 								uint64_t regionOffset = 0;
 
 								if (or.regionLength == 0)
